Type refactoring
