+++ src/semantic.rs
@@ -12,190 +12,17 @@
 //! Codegen is result of Semantic analyzer and contains prepared data tree
 //! of generated code for next step - compilation generated raw program code.
 use crate::ast::{self, GetLocation, GetName, MAX_PRIORITY_LEVEL_FOR_EXPRESSIONS};
-use crate::types::{error, SemanticStack, StructValue, TypeAttributes};
+use crate::types::block_state::BlockState;
+use crate::types::expression::{Expression, ExpressionResult, ExpressionResultValue};
+use crate::types::semantic::SemanticStack;
+use crate::types::types::{Type, TypeAttributes, TypeName};
 use crate::types::{
-    Binding, Constant, ConstantName, Expression, ExpressionResult, ExpressionResultValue, Function,
-    FunctionCall, FunctionName, FunctionStatement, InnerValueName, LabelName, LetBinding, Type,
-    TypeName, Value, ValueName,
+    error, Binding, Constant, ConstantName, Function, FunctionCall, FunctionName,
+    FunctionStatement, LabelName, LetBinding, StructValue, Value,
 };
 use std::cell::RefCell;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::rc::Rc;
-
-const IF_BEGIN: &str = "if_begin";
-const IF_END: &str = "if_end";
-const IF_ELSE: &str = "if_else";
-const LOOP_BEGIN: &str = "loop_begin";
-const LOOP_END: &str = "loop_end";
-
-/// # Block state
-/// - `values` - contains unique values map for current state but not unique
-///   for parent states. The map contains key-value: `value_name` (unique
-///   only for current state); and `Value` itself - value parameters.   
-/// - `inner_values_name` - is entity that represent inner value name - it
-///   can be different from `Value` name because it should be unique for all
-///   parent states. For example, of 3 values with name `x`, inner value
-///   name will be: [`x`, `x.0`, `x.1`]. It mean, inner value name can
-///   contain `value counter` as end of the name.
-/// - `labels` - labels set, for conditional operation. Unique for current
-///   and all paren states.
-/// - `last_register_number` - represent register counter for current and
-///   all parent states for `Codegen`. Register represented as `u64` and
-///   should be linearly incremented.
-/// - `manual_return` - flag indicated, that return was invoked from
-/// other state, for example: if-flow, loop-flow
-/// - `parent` - represent parent states.  
-#[derive(Debug)]
-pub struct BlockState {
-    /// State values
-    pub values: HashMap<ValueName, Value>,
-    /// Used to keep all names in the block state (and parent) as unique
-    pub inner_values_name: HashSet<InnerValueName>,
-    /// State labels for conditional operations
-    pub labels: HashSet<LabelName>,
-    /// Last register for unique register representation
-    pub last_register_number: u64,
-    /// Manual return from other states
-    pub manual_return: bool,
-    /// Parent state
-    pub parent: Option<Rc<RefCell<BlockState>>>,
-    /// children states
-    pub children: Vec<Rc<RefCell<BlockState>>>,
-    /// Semantic stack context for Block state
-    pub context: SemanticStack,
-}
-
-impl BlockState {
-    /// Init block state with optional `parent` state
-    pub fn new(parent: Option<Rc<RefCell<Self>>>) -> Self {
-        // Get values from parent
-        let (last_register_number, inner_values_name, labels, manual_return) =
-            parent.clone().map_or_else(
-                || (0, HashSet::new(), HashSet::new(), false),
-                |p| {
-                    let parent = p.borrow();
-                    (
-                        parent.last_register_number,
-                        parent.inner_values_name.clone(),
-                        parent.labels.clone(),
-                        parent.manual_return,
-                    )
-                },
-            );
-        Self {
-            values: HashMap::new(),
-            children: vec![],
-            inner_values_name,
-            labels,
-            last_register_number,
-            manual_return,
-            parent,
-            context: SemanticStack::new(),
-        }
-    }
-
-    fn set_child(&mut self, child: Rc<RefCell<BlockState>>) {
-        self.children.push(child);
-    }
-
-    /// Set value inner name to current state and parent states
-    fn set_inner_value_name(&mut self, name: &InnerValueName) {
-        self.inner_values_name.insert(name.clone());
-        if let Some(parent) = &self.parent {
-            parent.borrow_mut().set_inner_value_name(name);
-        }
-    }
-
-    /// Check is `inner_value_name` exist in current and parent states
-    fn is_inner_value_name_exist(&self, name: &InnerValueName) -> bool {
-        if self.inner_values_name.contains(name) {
-            return true;
-        } else if let Some(parent) = &self.parent {
-            return parent.borrow().is_inner_value_name_exist(name);
-        }
-        false
-    }
-
-    /// Get `Value` by value name from current state.
-    /// If not found on current state - recursively find in parent states.
-    fn get_value_name(&self, name: &ValueName) -> Option<Value> {
-        if let Some(val) = self.values.get(name) {
-            return Some(val.clone());
-        } else if let Some(parent) = &self.parent {
-            return parent.borrow().get_value_name(name);
-        }
-        None
-    }
-
-    /// Check is label name exist in current and parent states
-    fn is_label_name_exist(&self, name: &LabelName) -> bool {
-        if self.labels.contains(name) {
-            return true;
-        } else if let Some(parent) = &self.parent {
-            return parent.borrow().is_label_name_exist(name);
-        }
-        false
-    }
-
-    /// Set label name to current and all parent states
-    fn set_label_name(&mut self, name: &LabelName) {
-        self.labels.insert(name.clone());
-        if let Some(parent) = &self.parent {
-            parent.borrow_mut().set_label_name(name);
-        }
-    }
-
-    /// Set attribute counter - increment, if counter exist.
-    fn set_attr_counter(val: &str) -> String {
-        let val_attr: Vec<&str> = val.split('.').collect();
-        if val_attr.len() == 2 {
-            let i: u64 = val_attr[1].parse().expect("expect integer");
-            format!("{}.{:?}", val_attr[0], i + 1)
-        } else {
-            format!("{}.0", val_attr[0])
-        }
-    }
-
-    /// Get and set next label for condition operations
-    /// - If label doesn't exist in State - just insert to State and
-    ///   self return
-    /// - if label exists, get label counter
-    fn get_and_set_next_label(&mut self, label: &LabelName) -> LabelName {
-        // Check is label exists. If doesn't set it to State and return self
-        if !self.is_label_name_exist(label) {
-            self.set_label_name(label);
-            return label.clone();
-        }
-        // If label exists, split and get number of label counter
-        let name: LabelName = Self::set_attr_counter(&label.to_string()).into();
-        if self.is_label_name_exist(&name) {
-            self.get_and_set_next_label(&name)
-        } else {
-            self.set_label_name(&name);
-            name
-        }
-    }
-
-    /// Get next `inner_value_name` by name counter for current and
-    /// parent states. The `inner_value_name` should always be unique.
-    fn get_next_inner_name(&self, val: &InnerValueName) -> InnerValueName {
-        // Increment inner value name counter for shadowed variable
-        let name: InnerValueName = Self::set_attr_counter(&val.to_string()).into();
-        if self.is_inner_value_name_exist(&name) {
-            self.get_next_inner_name(&name)
-        } else {
-            name
-        }
-    }
-
-    /// Set return status flag for current and parent states
-    fn set_return(&mut self) {
-        self.manual_return = true;
-        if let Some(parent) = &self.parent {
-            parent.borrow_mut().set_return();
-        }
-    }
-}
 
 /// # Global State
 /// Global state can contains state of:
@@ -899,16 +726,16 @@ impl State {
         // Get labels name for if-begin, and if-end
         let label_if_begin = if_body_state
             .borrow_mut()
-            .get_and_set_next_label(&IF_BEGIN.to_string().into());
+            .get_and_set_next_label(&"if_begin".to_string().into());
         let label_if_else = if_body_state
             .borrow_mut()
-            .get_and_set_next_label(&IF_ELSE.to_string().into());
+            .get_and_set_next_label(&"if_else".to_string().into());
         // Set if-end label from previous context, if exist
         let label_if_end = label_end.map_or_else(
             || {
                 if_body_state
                     .borrow_mut()
-                    .get_and_set_next_label(&IF_END.to_string().into())
+                    .get_and_set_next_label(&"if_begin".to_string().into())
             },
             |label| label,
         );
@@ -1019,11 +846,11 @@ impl State {
         // Get labels name for loop-begin, and loop-end
         let label_loop_begin = loop_body_state
             .borrow_mut()
-            .get_and_set_next_label(&LOOP_BEGIN.to_string().into());
+            .get_and_set_next_label(&"loop_begin".to_string().into());
 
         let label_loop_end = loop_body_state
             .borrow_mut()
-            .get_and_set_next_label(&LOOP_END.to_string().into());
+            .get_and_set_next_label(&"loop_end".to_string().into());
 
         loop_body_state
             .borrow_mut()
